+++ CHANGELOG.md
@@ -11,6 +11,7 @@
 - Text: Remove deprecated prop __dangerouslyIncreaseLineHeight (#773)
 
 ### Patch
+- SelectList: Remove selected prop from the placeholder option tag for better React support (#759)
 
 </details>